fix: loan flow improves

+++ backend/src/main/java/mn_react/infrastructure/http/controllers/BookController.java
@@ -68,7 +68,7 @@ public class BookController {
             .content(content)
             .page(page)
             .size(size)
-            .totalElements(total)
+            .totalElements(((int) total))
             .build();
 
         return HttpResponse.ok(response);
