+++ backend/src/main/java/mn_react/infrastructure/http/dto/responses/PagedResponse.java
@@ -19,6 +19,6 @@ public class PagedResponse<T> {
     private List<T> content;
     private Integer page;
     private Integer size;
-    private Long totalElements;
+    private Integer totalElements;
     private Integer totalPages;
 }
