+++ backend/src/main/java/mn_react/infrastructure/persistence/jdbc/BookJdbcRepository.java
@@ -1,12 +1,12 @@
 package mn_react.infrastructure.persistence.jdbc;
 
+import java.util.List;
+
 import io.micronaut.data.annotation.Query;
 import io.micronaut.data.jdbc.annotation.JdbcRepository;
 import io.micronaut.data.model.query.builder.sql.Dialect;
 import io.micronaut.data.repository.CrudRepository;
 import mn_react.infrastructure.persistence.entity.BookEntity;
-
-import java.util.List;
 
 @JdbcRepository(dialect = Dialect.H2)
 public interface BookJdbcRepository extends CrudRepository<BookEntity, Long> {
@@ -37,5 +37,5 @@ public interface BookJdbcRepository extends CrudRepository<BookEntity, Long> {
         LIKE LOWER(CONCAT('%', :search, '%'))
     )
     """)
-    long countWithSearch(String search);
+    Integer countWithSearch(String search);
 }
